Add files via upload
created a function to calculate the distribution of ‘channeltype’ from the top 1000 records. The first focus is to need a data frame that contains only the top 1000 records from the 4000 original records, so we created a data frame “first_thousand” as the data frame contains the top 1000 records. Then we need to calculate the distribution of ‘channeltype’ from that record using a function. So, I defined the function as “dis_fun” with parameters ‘df’ (data frame) for allocating the data frame we created for the top 1000 records and ‘channeltype’ for allocating the column (channeltype). Then we print the data frame by grouping the ‘channeltype’ with the count of the channels. Then we called the functions with the parameters we already defined (data frame name and column name).